move all to the correct new place
